R525-028 Add support for editors with just a 'Run' button
Enable the running backend (in completely unsafe mode for now)
Code cleanup in editors.js.

For inline editors, do not require an AJAX request to get the
example base, assume everything is provided inline.

Add support for providing shadow files: these will not be displayed
in editors but will be transmitted to the backend.